채팅/메세지 기능 분리
