--- src/main/java/com/sparta/cucumber/chat/MeetRequestDto.java
@@ -1,10 +1,0 @@
-package com.sparta.cucumber.chat;
-
-import lombok.Data;
-
-@Data
-public class MeetRequestDto {
-    private Long id;
-    private Long commenterId;
-    private Long articleId;
-}
