+++ src/main/resources/static/src/index.js
@@ -1,4 +1,3 @@
-import moment from 'moment';
 import axios from 'axios';
 import $ from 'jquery'
 import '@popperjs/core'
@@ -8,8 +7,8 @@ import './css/main.css'
 import './kakao'
 import './aba5c3ead0';
 import SockJS from 'sockjs-client'
-import { Stomp } from '@stomp/stompjs'
-import { chatView, getArticles, logInView, registerView, setModal } from './view';
+import {Stomp} from '@stomp/stompjs'
+import {addComment, callComments, chatView, getArticles, logInView, registerView, setModal} from './view';
 
 let stompClient;
 let userId = null;
@@ -17,7 +16,7 @@ Kakao.init("e1289217c77f4f46dc511544f119d102");
 
 (function setHeader() {
     let token = localStorage.getItem("token");
-    axios.defaults.headers.common = { Authorization: `Bearer ${token}` }
+    axios.defaults.headers.common = {Authorization: `Bearer ${token}`}
 })()
 
 function connect() {
@@ -200,35 +199,6 @@ export function writeComment(idx) {
             console.log(error);
         });
 }
-
-function callComments(idx) {
-    axios
-        .get(`${API_URL}/api/comments/${idx}`)
-        .then((response) => {
-            let { data } = response
-            data.forEach((comment) => {
-                addComment(idx, comment);
-            })
-        })
-}
-
-export function addComment(idx, data) {
-    userId = parseInt(localStorage.getItem("userId"));
-    let { id, content, createdAt, user } = data;
-    console.log(userId)
-    $(`#comment-list-${idx}`).append(`
-    <li href="#" class="list-group-item list-group-item-action">
-    <div class="d-flex w-100 justify-content-between">
-      <small class="mb-1"><small class="mb-1 tit">${user.name}</small>
-      ${moment(createdAt).fromNow()}</small>
-      ${userId === user.id
-            ? `<button type="button" class="btn-close small" aria-label="remove" onclick="app.removeComment(${idx}, ${id})"></button>`
-            : `<button onclick="app.letsMeet(${idx}, ${user.id})" class="badge bg-success rounded-pill">chat</button>`}
-    </div>
-    <p class="mb-1">${content}</small>
-  </li>`);
-}
-
 
 export function letsMeet(idx, userId) {
     const body = {
