+++ src/main/resources/static/src/view.js
@@ -1,3 +1,9 @@
+import moment from 'moment';
+import axios from 'axios';
+import $ from 'jquery'
+
+let userId = null;
+
 const getArticles = () => {
     let div = document.createElement("div");
     div.className = "card-deck";
@@ -193,4 +199,33 @@ function chatView() {
   </div>`
 }
 
-export {getArticles, setModal, registerView, logInView, chatView};
+function addComment(idx, data) {
+    userId = parseInt(localStorage.getItem("userId"));
+    let {id, content, createdAt, user} = data;
+    console.log(userId)
+    $(`#comment-list-${idx}`).append(`
+  <li href="#" class="list-group-item list-group-item-action">
+  <div class="d-flex w-100 justify-content-between">
+    <small class="mb-1"><small class="mb-1 tit">${user.name}</small>
+    ${moment(createdAt).fromNow()}</small>
+    ${userId === user.id
+        ? `<button type="button" class="btn-close small" aria-label="remove" onclick="app.removeComment(${idx}, ${id})"></button>`
+        : `<button onclick="app.letsMeet(${idx}, ${user.id})" class="badge bg-success rounded-pill">chat</button>`}
+  </div>
+  <p class="mb-1">${content}</small>
+</li>`);
+}
+
+
+function callComments(idx) {
+    axios
+        .get(`${API_URL}/api/comments/${idx}`)
+        .then((response) => {
+            let {data} = response
+            data.forEach((comment) => {
+                addComment(idx, comment);
+            })
+        })
+}
+
+export {getArticles, setModal, registerView, logInView, chatView, addComment, callComments};
